+++ components/marketplace/utils/ratinhos.ts
@@ -70,6 +70,258 @@ export const ratinhos = [
     priceMTH: randomNumbers(),
     image: image,
   },
+  {
+    id: 7,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 8,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 9,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 10,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 11,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 12,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 13,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 14,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 15,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 16,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 17,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 18,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 19,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 20,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 21,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 22,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 23,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 24,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 25,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 26,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 27,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 28,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 29,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 30,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 31,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 32,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 33,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
+  {
+    id: 34,
+    name: 'Legendary Mice',
+    date_posted: randomDate(),
+    rarity: 'legendary',
+    priceUsd: randomNumbers(),
+    priceMTH: randomNumbers(),
+    image: image,
+  },
 ];
 
 export interface Rats {
@@ -77,6 +329,9 @@ export interface Rats {
   name: string;
   date_posted: Date;
   rarity: string;
+  priceUsd: number;
+  priceMTH: number;
+  image: string;
 }
 
 export interface Props {
